requirements.txt
requests==2.31.0
PyGithub==2.1.1
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml==4.9.3